Cirrus: Temp. workaround missing imgprune image
The 'image_prune' task only runs on master, post-merge and
is currently failing for all builds.  This is because it
references a non-existent image.  The person with access
to add/enable this image is on PTO.  Fix this by temporarily
using a hand-built image until an automatic build can be added.

Signed-off-by: Chris Evich <cevich@redhat.com>