Test that restored container does not depend on the original container
In the restore from external checkpoint archive test, the second restore
using a new name and ID is now done first to ensure that nothing in the
restored container depends on the original container.

Test has been adapted to catch errors like the one fixed with the
previous commit to adapt ConmonPidFile for restored containers.

Signed-off-by: Adrian Reber <areber@redhat.com>