Adjust name of Podman CNI network bridge
Both Podman and CRI-O set up CNI bridges with the name 'cni0'. If
both our CNI conflist and the CRI-O conflist are installed,
whoever runs first will win - that is, they will configure the
bridge, and everyone will use it. Problem: the CRI-O CNI config
conflicts with ours and results in containers with no networking.
Solution: rename our bridge so we don't conflict with CRI-O.

At the same time, hit our IPAM configuration. The current version
was an older format in danger of deprecation. The new format is
documented at [1].

Finally, fix indentation for the entire file.

[1] https://github.com/containernetworking/plugins/tree/master/plugins/ipam/host-local#example-configurations

Signed-off-by: Matthew Heon <matthew.heon@pm.me>